+++ pysat/tests/test_constellation.py
@@ -3,9 +3,9 @@
 # Full author list can be found in .zenodo.json file
 # DOI:10.5281/zenodo.1199703
 # ----------------------------------------------------------------------------
+"""Unit tests for the `constellation` class and methods."""
 
 import datetime as dt
-from io import StringIO
 import logging
 import pandas as pds
 import pytest
@@ -16,15 +16,15 @@ from pysat.tests.registration_test_class import TestWithRegistration
 
 
 class TestConstellationInitReg(TestWithRegistration):
-    """Test the Constellation class initialization with registered Instruments.
-    """
+    """Unit tests for the Constellation class with registered Instruments."""
+
     @pytest.mark.parametrize("ikeys, ivals, ilen",
                              [(["platforms", "tags"], [["platname1"], [""]], 2),
                               (["names", "tags"], [["name2"], [""]], 2),
                               (["names"], [["name1", "name2"]], 15)])
     def test_construct_constellation(self, ikeys, ivals, ilen):
-        """Construct a Constellation with good input
-        """
+        """Test construction of a Constellation with good input."""
+
         # Register fake Instrument modules
         pysat.utils.registry.register(self.module_names)
 
@@ -38,8 +38,8 @@ class TestConstellationInitReg(TestWithRegistration):
         return
 
     def test_all_bad_construct_constellation(self):
-        """Test raises ValueError when all inputs are unregistered
-        """
+        """Test raises ValueError when all inputs are unregistered."""
+
         # Register fake Instrument modules
         pysat.utils.registry.register(self.module_names)
 
@@ -50,55 +50,53 @@ class TestConstellationInitReg(TestWithRegistration):
         assert str(verr).find("no registered packages match input") >= 0
         return
 
-    def test_some_bad_construct_constellation(self):
-        """Test partial load and log warning when some inputs are unregistered
-        """
-        # Initialize logging
-        log_capture = StringIO()
-        pysat.logger.addHandler(logging.StreamHandler(log_capture))
-        pysat.logger.setLevel(logging.WARNING)
+    def test_some_bad_construct_constellation(self, caplog):
+        """Test partial load and log warning if some inputs are unregistered."""
 
         # Register fake Instrument modules
         pysat.utils.registry.register(self.module_names)
 
         # Load the Constellation and capture log output
-        const = pysat.Constellation(platforms=['Executor', 'platname1'],
-                                    tags=[''])
-        log_out = log_capture.getvalue()
+        with caplog.at_level(logging.WARNING, logger='pysat'):
+            const = pysat.Constellation(platforms=['Executor', 'platname1'],
+                                        tags=[''])
 
         # Test the partial Constellation initialization
         assert len(const.instruments) == 2
 
         # Test the log warning
-        assert log_out.find("unable to load some platforms") >= 0
+        assert caplog.text.find("unable to load some platforms") >= 0
 
-        del log_capture, log_out, const
+        del const
         return
 
 
-class TestConstellationInit:
+class TestConstellationInit(object):
     """Test the Constellation class."""
+
     def setup(self):
-        """Create instruments and a constellation for each test
-        """
+        """Set up the unit test environment for each method."""
+
         self.instruments = constellations.single_test.instruments
         self.in_kwargs = {"instruments": self.instruments,
                           "const_module": constellations.single_test}
         self.const = None
         self.ref_time = pysat.instruments.pysat_testing._test_dates['']['']
+        return
 
     def teardown(self):
-        """Clean up after each test
-        """
+        """Clean up the unit test environment after each method."""
+
         del self.const, self.instruments, self.in_kwargs, self.ref_time
+        return
 
     @pytest.mark.parametrize("ikey,ival,ilen",
                              [("const_module", None, 1),
                               ("instruments", None, 1),
                               (None, None, 2)])
     def test_construct_constellation(self, ikey, ival, ilen):
-        """Construct a Constellation with good input
-        """
+        """Test construction of a Constellation with good input."""
+
         if ikey is not None:
             self.in_kwargs[ikey] = ival
         self.const = pysat.Constellation(**self.in_kwargs)
@@ -106,8 +104,8 @@ class TestConstellationInit:
         return
 
     def test_init_constellation_bad_inst_module(self):
-        """Test Constellation raises AttributeError with bad inst_module input.
-        """
+        """Test Constellation raises AttributeError with bad inst_module."""
+
         with pytest.raises(AttributeError) as aerr:
             pysat.Constellation(const_module=self.instruments)
 
@@ -115,8 +113,8 @@ class TestConstellationInit:
         return
 
     def test_construct_raises_noniterable_error(self):
-        """Attempt to construct a Constellation by const_module and list
-        """
+        """Test error raised when Constellation non iterable."""
+
         with pytest.raises(ValueError) as verr:
             self.const = pysat.Constellation(instruments=self.instruments[0])
 
@@ -124,15 +122,15 @@ class TestConstellationInit:
         return
 
     def test_construct_null(self):
-        """Attempt to construct a Constellation with no arguments
-        """
+        """Test that a Constellation constructed without arguments is empty."""
+
         self.const = pysat.Constellation()
         assert len(self.const.instruments) == 0
         return
 
     def test_getitem(self):
-        """Test Constellation iteration through instruments attribute
-        """
+        """Test Constellation iteration through instruments attribute."""
+
         self.in_kwargs['const_module'] = None
         self.const = pysat.Constellation(**self.in_kwargs)
         tst_get_inst = self.const[:]
@@ -140,8 +138,8 @@ class TestConstellationInit:
         return
 
     def test_repr_w_inst(self):
-        """Test Constellation string output with instruments loaded
-        """
+        """Test Constellation string output with instruments loaded."""
+
         self.in_kwargs['const_module'] = None
         self.const = pysat.Constellation(**self.in_kwargs)
         out_str = self.const.__repr__()
@@ -150,8 +148,8 @@ class TestConstellationInit:
         return
 
     def test_str_w_inst(self):
-        """Test Constellation string output with instruments loaded
-        """
+        """Test Constellation string output with instruments loaded."""
+
         self.in_kwargs['const_module'] = None
         self.const = pysat.Constellation(**self.in_kwargs)
         out_str = self.const.__str__()
@@ -161,8 +159,8 @@ class TestConstellationInit:
         return
 
     def test_str_wo_inst(self):
-        """Test Constellation string output without instruments.
-        """
+        """Test Constellation string output without instruments."""
+
         self.const = pysat.Constellation()
         out_str = self.const.__str__()
 
@@ -174,6 +172,7 @@ class TestConstellationInit:
                                                    (False, "Full")])
     def test_str_with_data(self, common_index, cstr):
         """Test Constellation string output with loaded data."""
+
         self.in_kwargs["common_index"] = common_index
         self.const = pysat.Constellation(**self.in_kwargs)
         self.const.load(date=self.ref_time)
@@ -184,8 +183,8 @@ class TestConstellationInit:
         return
 
     def test_single_attachment_of_custom_function(self):
-        """Test successful attachment of custom function
-        """
+        """Test successful attachment of custom function."""
+
         # Define a custom function
         def double_mlt(inst):
             dmlt = 2.0 * inst.data.mlt
@@ -208,11 +207,12 @@ class TestConstellationInit:
         return
 
 
-class TestConstellationFunc:
+class TestConstellationFunc(object):
     """Test the Constellation class attributes and methods."""
+
     def setup(self):
-        """Create instruments and a constellation for each test
-        """
+        """Set up the unit test environment for each method."""
+
         self.inst = list(constellations.testing.instruments)
         self.const = pysat.Constellation(instruments=self.inst)
         self.ref_time = pysat.instruments.pysat_testing._test_dates['']['']
@@ -220,11 +220,13 @@ class TestConstellationFunc:
                       "bounds", "empty", "empty_partial", "index_res",
                       "common_index", "date", "yr", "doy", "yesterday", "today",
                       "tomorrow", "variables"]
+        return
 
     def teardown(self):
-        """Clean up after each test
-        """
+        """Clean up the unit test environment after each method."""
+
         del self.inst, self.const, self.ref_time, self.attrs
+        return
 
     def test_has_required_attrs(self):
         """Ensure the instrument has all required attributes present."""
@@ -241,7 +243,7 @@ class TestConstellationFunc:
         return
 
     def test_bounds_passthrough(self):
-        """Ensure bounds are applied to each instrument within Constellation"""
+        """Ensure bounds are applied to each instrument within Constellation."""
 
         # Set bounds
         stop_date = self.ref_time + dt.timedelta(days=365)
@@ -256,7 +258,8 @@ class TestConstellationFunc:
         return
 
     def test_empty_data_index(self):
-        """ Test the empty index attribute."""
+        """Test the empty index attribute."""
+
         # Test the attribute with no loaded data
         assert isinstance(self.const.index, pds.Index)
         assert len(self.const.index) == 0
@@ -264,22 +267,26 @@ class TestConstellationFunc:
 
     def test_empty_data_date(self):
         """Test the date property when no data is loaded."""
+
         assert self.const.date is None
         return
 
     def test_empty_variables(self):
         """Test the variables property when no data is loaded."""
+
         assert len(self.const.variables) == 0
         return
 
     def test_empty_flag_data_empty(self):
-        """ Test the status of the empty flag for unloaded data."""
+        """Test the status of the empty flag for unloaded data."""
+
         assert self.const.empty
         assert self.const.empty_partial
         return
 
     def test_empty_flag_data_empty_partial_load(self):
-        """ Test the status of the empty flag for partially loaded data."""
+        """Test the status of the empty flag for partially loaded data."""
+
         # Load only one instrument and test the status flag
         self.const.instruments[0].load(date=self.ref_time)
         assert self.const.empty_partial
@@ -288,13 +295,15 @@ class TestConstellationFunc:
 
     def test_empty_flag_data_not_empty_partial_load(self):
         """Test the alt status of the empty flag for partially loaded data."""
+
         # Load only one instrument and test the status flag for alternate flag
         self.const.instruments[0].load(date=self.ref_time)
         assert not self.const._empty(all_inst=False)
         return
 
     def test_empty_flag_data_not_empty(self):
-        """ Test the status of the empty flag for loaded data."""
+        """Test the status of the empty flag for loaded data."""
+
         # Load data and test the status flag
         self.const.load(date=self.ref_time)
         assert not self.const.empty
@@ -303,7 +312,8 @@ class TestConstellationFunc:
     @pytest.mark.parametrize("ikwarg", [{"common_index": False},
                                         {"index_res": 60.0}])
     def test_full_data_index(self, ikwarg):
-        """ Test the empty index attribute."""
+        """Test the empty index attribute."""
+
         # Test the attribute with loaded data
         self.const = pysat.Constellation(instruments=self.inst, **ikwarg)
         self.const.load(date=self.ref_time)
@@ -316,8 +326,8 @@ class TestConstellationFunc:
         return
 
     def test_today_yesterday_and_tomorrow(self):
-        """ Test the correct instantiation of yesterday/today/tomorrow dates
-        """
+        """Test the correct instantiation of yesterday/today/tomorrow dates."""
+
         for cinst in self.const.instruments:
             assert cinst.today() == self.const.today()
             assert cinst.yesterday() == self.const.yesterday()
@@ -326,6 +336,7 @@ class TestConstellationFunc:
 
     def test_full_data_date(self):
         """Test the date property when no data is loaded."""
+
         # Test the attribute with loaded data
         self.const.load(date=self.ref_time)
 
@@ -334,6 +345,7 @@ class TestConstellationFunc:
 
     def test_full_variables(self):
         """Test the variables property when no data is loaded."""
+
         # Test the attribute with loaded data
         self.const.load(date=self.ref_time)
 
@@ -344,6 +356,7 @@ class TestConstellationFunc:
 
     def test_download(self):
         """Check that instruments are downloadable."""
+
         self.const.download(self.ref_time, self.ref_time)
         for inst in self.const.instruments:
             assert len(inst.files.files) > 0
@@ -351,6 +364,7 @@ class TestConstellationFunc:
 
     def test_get_unique_attr_vals_bad_attr(self):
         """Test raises AttributeError for bad input value."""
+
         with pytest.raises(AttributeError) as aerr:
             self.const._get_unique_attr_vals('not_an_attr')
 
@@ -359,6 +373,7 @@ class TestConstellationFunc:
 
     def test_get_unique_attr_vals_bad_type(self):
         """Test raises AttributeError for bad input attribute type."""
+
         with pytest.raises(TypeError) as terr:
             self.const._get_unique_attr_vals('empty')