+++ pysat/tests/test_methods_general.py
@@ -1,3 +1,5 @@
+"""Unit tests for the general instrument methods."""
+
 import datetime as dt
 from os import path
 import pandas as pds
@@ -7,18 +9,23 @@ import pysat
 from pysat.instruments.methods import general as gen
 
 
-class TestGenMethods():
+class TestListFiles(object):
+    """Unit tests for `pysat.instrument.methods.general.list_files`."""
 
     def setup(self):
-        """Runs before every method to create a clean testing setup."""
+        """Set up the unit test environment for each method."""
+
         fname = 'fake_data_{year:04d}{month:02d}{day:02d}_v05.cdf'
         self.kwargs = {'tag': '', 'inst_id': '', 'data_path': '/fake/path/',
                        'format_str': None,
                        'supported_tags': {'': {'': fname}}}
+        return
 
     def teardown(self):
-        """Runs after every method to clean up previous testing."""
+        """Clean up the unit test environment after each method."""
+
         del self.kwargs
+        return
 
     @pytest.mark.parametrize("bad_key,bad_val,err_msg",
                              [("data_path", None,
@@ -26,25 +33,30 @@ class TestGenMethods():
                               ("tag", "badval", "Unknown inst_id or tag"),
                               ("inst_id", "badval", "Unknown inst_id or tag")])
     def test_bad_kwarg_list_files(self, bad_key, bad_val, err_msg):
+        """Test that bad kwargs raise an appropriate error."""
+
         self.kwargs[bad_key] = bad_val
         with pytest.raises(ValueError) as excinfo:
             gen.list_files(**self.kwargs)
         assert str(excinfo.value).find(err_msg) >= 0
+        return
 
 
-class TestFileCadence():
+class TestFileCadence(object):
+    """Unit tests for `instruments.methods.general.is_daily_file_cadence`."""
 
     @pytest.mark.parametrize("time_kwarg, time_val, is_daily",
                              [("microseconds", 1, True), ("seconds", 1, True),
                               ("minutes", 1, True), ("hours", 1, True),
                               ("days", 1, True), ("days", 2, False)])
     def test_datetime_file_cadence(self, time_kwarg, time_val, is_daily):
-        """ Test is_daily_file_cadence with dt.datetime input
-        """
+        """Test `is_daily_file_cadence` with dt.datetime input."""
+
         in_time = dt.timedelta(**{time_kwarg: time_val})
         check_daily = gen.is_daily_file_cadence(in_time)
 
         assert check_daily == is_daily
+        return
 
     @pytest.mark.parametrize("time_kwarg, time_val, is_daily",
                              [("microseconds", 1, True), ("seconds", 1, True),
@@ -52,33 +64,46 @@ class TestFileCadence():
                               ("days", 1, True), ("days", 2, False),
                               ("months", 1, False), ("years", 1, False)])
     def test_dateoffset_file_cadence(self, time_kwarg, time_val, is_daily):
-        """ Test is_daily_file_cadence with pds.DateOffset input
-        """
+        """Test `is_daily_file_cadence` with pds.DateOffset input."""
+
         in_time = pds.DateOffset(**{time_kwarg: time_val})
         check_daily = gen.is_daily_file_cadence(in_time)
 
         assert check_daily == is_daily
+        return
 
 
-class TestRemoveLeadText():
+class TestRemoveLeadText(object):
+    """Unit tests for `pysat.instrument.methods.general.remove_leading_text`."""
+
     def setup(self):
-        """Runs before every method to create a clean testing setup."""
+        """Set up the unit test environment for each method."""
+
         # Load a test instrument
         self.testInst = pysat.Instrument('pysat', 'testing', num_samples=12,
                                          clean_level='clean')
         self.testInst.load(2009, 1)
         self.npts = len(self.testInst['uts'])
+        return
 
     def teardown(self):
-        """Runs after every method to clean up previous testing."""
+        """Clean up the unit test environment after each method."""
+
         del self.testInst, self.npts
+        return
 
     def test_remove_prefix_w_bad_target(self):
+        """Test that a bad target in `remove_leading_text` raises an error."""
+
         self.testInst['ICON_L27_Blurp'] = self.testInst['dummy1']
-        with pytest.raises(ValueError):
+        with pytest.raises(ValueError) as verr:
             gen.remove_leading_text(self.testInst, target=17.5)
+        assert str(verr).find('target must be a string or list of strings') >= 0
+        return
 
     def test_remove_names_wo_target(self):
+        """Test that an unspecified target leaves variable names unchanged."""
+
         self.testInst['ICON_L27_Blurp'] = self.testInst['dummy1']
         gen.remove_leading_text(self.testInst)
 
@@ -87,8 +112,11 @@ class TestRemoveLeadText():
 
         # Check other names untouched
         assert len(self.testInst['dummy1']) == self.npts
+        return
 
     def test_remove_names_w_target(self):
+        """Test that only names with the target prefix are changed."""
+
         self.testInst['ICON_L27_Blurp'] = self.testInst['dummy1']
         gen.remove_leading_text(self.testInst, target='ICON_L27')
 
@@ -100,8 +128,11 @@ class TestRemoveLeadText():
 
         # Check prepended text removed from metadata
         assert '_Blurp' in self.testInst.meta.keys()
+        return
 
     def test_remove_names_w_target_list(self):
+        """Test that multiple targets can be removed via a list."""
+
         self.testInst['ICON_L27_Blurp'] = self.testInst['dummy1']
         self.testInst['ICON_L23_Bloop'] = self.testInst['dummy1']
         gen.remove_leading_text(self.testInst,
@@ -117,23 +148,37 @@ class TestRemoveLeadText():
         # Check prepended text removed from metadata
         assert '_Blurp' in self.testInst.meta.keys()
         assert 'loop' in self.testInst.meta.keys()
+        return
 
 
 class TestRemoveLeadTextXarray(TestRemoveLeadText):
+    """Unit tests for `pysat.instrument.methods.general.remove_leading_text`.
+
+    Note
+    ----
+    Includes additional xarray-specific tests.
+    """
+
     def setup(self):
-        """Runs before every method to create a clean testing setup."""
+        """Set up the unit test environment for each method."""
+
         # Load a test instrument
         self.testInst = pysat.Instrument('pysat', 'testing2d_xarray',
                                          num_samples=12,
                                          clean_level='clean')
         self.testInst.load(2009, 1)
         self.npts = len(self.testInst['uts'])
+        return
 
     def teardown(self):
-        """Runs after every method to clean up previous testing."""
+        """Clean up the unit test environment after each method."""
+
         del self.testInst, self.npts
+        return
 
     def test_remove_2D_names_w_target(self):
+        """Test that 2D variables are changed appropriately."""
+
         gen.remove_leading_text(self.testInst, target='variable')
 
         # Check prepended text removed from variables
@@ -142,8 +187,11 @@ class TestRemoveLeadTextXarray(TestRemoveLeadText):
 
         # Check prepended text removed from metadata
         assert '_profiles' in self.testInst.meta.keys()
+        return
 
     def test_remove_2D_names_w_target_list(self):
+        """Test that 2D variables can be modified with a target list."""
+
         gen.remove_leading_text(self.testInst,
                                 target=['variable', 'im'])
 
@@ -155,21 +203,28 @@ class TestRemoveLeadTextXarray(TestRemoveLeadText):
         # Check prepended text removed from metadata
         assert '_profiles' in self.testInst.meta.keys()
         assert 'ages' in self.testInst.meta.keys()
+        return
 
 
-class TestLoadCSVData():
+class TestLoadCSVData(object):
+    """Unit tests for `pysat.instrument.methods.general.load_csv_data`."""
+
     def setup(self):
-        """Runs before every method to create a clean testing setup."""
+        """Set up the unit test environment for each method."""
+
         # Load a test instrument
         self.csv_file = path.join(path.abspath(path.dirname(__file__)),
                                   'test_data', 'sw', 'kp', 'recent',
                                   'kp_recent_2019-03-18.txt')
         self.data_cols = ['mid_lat_Kp', 'high_lat_Kp', 'Kp']
         self.data = None
+        return
 
     def teardown(self):
-        """Runs after every method to clean up previous testing."""
+        """Clean up the unit test environment after each method."""
+
         del self.csv_file, self.data, self.data_cols
+        return
 
     def eval_data_cols(self):
         """Evaluate the data columns in the output."""