+++ pysat/utils/_core.py
@@ -144,7 +144,7 @@ def listify(iterable):
 
 
 def load_netcdf4(fnames=None, strict_meta=False, file_format=None,
-                 epoch_name='Epoch', pandas_format=True,
+                 epoch_name='Epoch', pandas_format=True, decode_timedelta=False,
                  labels={'units': ('units', str), 'name': ('long_name', str),
                          'notes': ('notes', str), 'desc': ('desc', str),
                          'min_val': ('value_min', np.float64),
@@ -172,6 +172,10 @@ def load_netcdf4(fnames=None, strict_meta=False, file_format=None,
     pandas_format : bool
         Flag specifying if data is stored in a pandas DataFrame (True) or
         xarray Dataset (False). (default=False)
+    decode_timedelta : bool
+        Used for xarray datasets.  If True, variables with unit attributes that
+        are 'timelike' ('hours', 'minutes', etc) are converted to
+        `np.timedelta64`. (default=False)
     labels : dict
         Dict where keys are the label attribute names and the values are tuples
         that have the label values and value types in that order.
@@ -210,6 +214,9 @@ def load_netcdf4(fnames=None, strict_meta=False, file_format=None,
                                             epoch_name=epoch_name,
                                             pandas_format=pandas_format,
                                             labels=labels)
+
+        # Close any open links to file through xarray.
+        out.close()
 
     return out, meta